fix(locks): call the fs2 unlock trait method explicitly, and clear clippy
std gained an inherent File::unlock in 1.89 that shadows fs2::FileExt
above the 1.83 MSRV, so the audit spool and taint store would silently
switch implementation with the toolchain; name the trait method instead.

Also reattach a summary doc line that had drifted off scan_single_file,
close a doc list continuation, build two test policies as literals, and
take the repo-standard allow on the per-scan manifest parser.

+++ crates/tirith-core/src/checkpoint.rs
@@ -1712,7 +1712,7 @@ fn available_bytes(path: &Path) -> Option<u64> {
     }
     // SAFETY: successful statvfs initialized the struct.
     let stat = unsafe { stat.assume_init() };
-    Some((stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64))
+    Some((stat.f_bavail as u64).saturating_mul(stat.f_frsize))
 }
 
 /// Backup a single file to the checkpoint files directory.
